Creating event handlers structure
