+++ init.lua
@@ -1,3 +1,4 @@
 require "config.lazy"
-require 'options'
-require 'keymaps'
+require "options"
+require "keymaps"
+require "event_handers"
