+++ lua/custom/plugins/ident-blankline.lua
@@ -1,4 +1,3 @@
---- @type LazyPluginSpec
 return {
     "lukas-reineke/indent-blankline.nvim",
     -- main = "ibl",
