+++ lua/custom/plugins/neogit.lua
@@ -1,15 +1,13 @@
 return {
-    {
-        "NeogitOrg/neogit",
-        dependencies = {
-            "nvim-lua/plenary.nvim",  -- required
-            "sindrets/diffview.nvim", -- optional - Diff integration
+    "NeogitOrg/neogit",
+    dependencies = {
+        "nvim-lua/plenary.nvim",  -- required
+        "sindrets/diffview.nvim", -- optional - Diff integration
 
-            -- Only one of these is needed.
-            "nvim-telescope/telescope.nvim", -- optional
-            "ibhagwan/fzf-lua",              -- optional
-            "nvim-mini/mini.pick",           -- optional
-            "folke/snacks.nvim",             -- optional
-        },
-    }
+        -- Only one of these is needed.
+        "nvim-telescope/telescope.nvim", -- optional
+        "ibhagwan/fzf-lua",              -- optional
+        "nvim-mini/mini.pick",           -- optional
+        "folke/snacks.nvim",             -- optional
+    },
 }
