+++ lua/custom/plugins/snacks.lua
@@ -1,34 +1,32 @@
 return {
-    {
-        "folke/snacks.nvim",
-        priority = 1000,
-        lazy = false,
-        opts = {
-            bigfile = { enabled = true },
-            dashboard = {
-                formats = {
-                    key = function(item)
-                        return { { "[", hl = "special" }, { item.key, hl = "key" }, { "]", hl = "special" } }
-                    end,
-                },
-                sections = {
-                    { title = "Recent",         padding = 1 },
-                    { section = "recent_files", padding = 1 },
-                    { title = "Sessions",       padding = 1 },
-                    { section = "projects",     padding = 1 },
-                    { section = "keys",         padding = 1 },
-                },
+    "folke/snacks.nvim",
+    priority = 1000,
+    lazy = false,
+    opts = {
+        bigfile = { enabled = true },
+        dashboard = {
+            formats = {
+                key = function(item)
+                    return { { "[", hl = "special" }, { item.key, hl = "key" }, { "]", hl = "special" } }
+                end,
             },
-            explorer = { enabled = true },
-            indent = { enabled = true },
-            input = { enabled = true },
-            picker = { enabled = true },
-            notifier = { enabled = true },
-            quickfile = { enabled = true },
-            scope = { enabled = true },
-            scroll = { enabled = true },
-            statuscolumn = { enabled = true },
-            words = { enabled = true },
+            sections = {
+                { title = "Recent",         padding = 1 },
+                { section = "recent_files", padding = 1 },
+                { title = "Sessions",       padding = 1 },
+                { section = "projects",     padding = 1 },
+                { section = "keys",         padding = 1 },
+            },
         },
-    }
+        explorer = { enabled = true },
+        indent = { enabled = true },
+        input = { enabled = true },
+        picker = { enabled = true },
+        notifier = { enabled = true },
+        quickfile = { enabled = true },
+        scope = { enabled = true },
+        scroll = { enabled = true },
+        statuscolumn = { enabled = true },
+        words = { enabled = true },
+    },
 }
