+++ lua/custom/plugins/web-devicons.lua
@@ -1,3 +1,1 @@
-return {
-    { "nvim-tree/nvim-web-devicons", opts = {} },
-}
+return { "nvim-tree/nvim-web-devicons", opts = {} }
