+++ lua/event_handers.lua
@@ -1,0 +1,1 @@
+require "scripts.buffer.event_bufdelete"
